Fix PR comments

+++ scripts/build.sh
@@ -6,9 +6,46 @@ WORKSPACE=$(pwd)
 echo "WORKSPACE is $WORKSPACE"
 
 CYCLONEDDS_INSTALL_PREFIX=$WORKSPACE/build/install/
+CLEAN_BUILD=false
 
-rm -rf build/
-mkdir build
+usage() {
+	echo ""
+	echo "Helper script to build cyclonedds"
+	echo "Usage : "
+	echo "<workspace>/scripts/build.sh [options]"
+	echo "Supported options are "
+	echo "clean 		: remove build folder (Default : keep build folder)"
+	echo ""
+}
+
+parse_args() {
+	for arg in "$@"
+	do
+		case $arg in
+		    -h | --help)
+			usage
+			exit 0
+			;;
+			clean)
+			CLEAN_BUILD=true
+			;;
+			*)
+			echo "ERROR : unknown parameter $arg"
+			usage
+			exit 1
+			;;
+		esac
+	done
+}
+
+parse_args $@
+
+if [ $CLEAN_BUILD == true ]
+then
+    rm -rf build/
+fi
+
+mkdir -p build
 cd build
 cmake -DCMAKE_INSTALL_PREFIX=$CYCLONEDDS_INSTALL_PREFIX -DBUILD_IDLC=ON ..
 cmake --build .
